test: split deferred recolor LRU fill into its own spec
Idle-fill and defer-window cases now live beside the WebGL parity file instead of inside it.
