+++ tests/tests.js
@@ -14,6 +14,7 @@ import "./canvas/palette-recolor-cache_spec.ts";
 import "./canvas/palette-recolor-merge_spec.ts";
 import "./canvas/palette-recolor-mode_spec.ts";
 import "./canvas/palette-recolor-webgl_spec.js";
+import "./canvas/palette-recolor-deferred_spec.ts";
 import "./canvas/palette-recolor-isolation_spec.ts";
 import "./canvas/preview-animation_spec.js";
 import "./canvas/preview-canvas_spec.js";
